Testing an idea about the BSD Conjecture
Using an Explicit Formula for the L-Function of an Elliptic Curve, it
appears that we can assign to each Elliptic Curve a Chebyshev-type
function constructed from the modular coefficients a_q=q+1-N_q. If the
Explicit Formula is correct, and assuming the General Riemann
Hypothesis, then as the input x increase, this function  should look
more and more line rx, where r is the analytic rank of the L-Function.
There may be a possible nontrivial contribution from the nontrivial
zeros, even assuming the GRH. This is to test 1.) The formula in general
and 2.) Whether or not there  is a nontrivial contribution.

Counting points on the elliptic curve is the hardest thing. There is the
Schoof algorithm to do it, but it would require a larger framework to
work within. I originally tried it by just seeing if every point mod p
was on the curve, but later realized I could just use Euler's Criterion
to see if x^3+Ax+B was a square or not mod p, which would be enough.

I've tried it for two curves  up to x~10,000 and the values appear to
float around the BSD-predicted algebraic rank, but it is not conclusive
enough to rule out some kind of error.